feat: 모달 파일 네이밍, 렌더링에서 이메일 뺌
